Feature

Former-commit-id: d75788c33e1ae932e4c17374505e795fb0594684
Former-commit-id: 86f20e596d830754716403ae216a0365894e3fff

+++ core/domain/member/member.go
@@ -544,7 +544,7 @@ func (m *memberImpl) memberInit() error {
 	regPresent := m.registryRepo.Get(registry.MemberPresentIntegralNumOfRegister).IntValue()
 	if regPresent > 0 {
 		go m.GetAccount().Charge(member.AccountIntegral, "新会员注册赠送积分",
-			float32(regPresent), "-", "sys")
+			regPresent, "-", "sys")
 	}
 	return nil
 }
